\name{adf-package}
\alias{adf-package}
\alias{adf}
\docType{package}
\title{
\packageTitle{adf}
}
\description{
\packageDescription{adf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{adf}
\packageIndices{adf}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{adf}

Maintainer: \packageMaintainer{adf}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
