% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFactorLevels.r
\name{allFactorLevels}
\alias{allFactorLevels}
\title{Get All Factor Levels of a (Abstract) Data Frame}
\usage{
allFactorLevels(x)
}
\arguments{
\item{x}{(abstract) data frame}
}
\description{
Get all factor levels of a (possibly abstract) data frame.
}
