% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.apply.function.r
\name{default.apply.function}
\alias{default.apply.function}
\title{Default Method for Applying a Function to an Abstrct Data Frame}
\usage{
default.apply.function(x, FUN2, params)
}
\arguments{
\item{x}{an abstract data frame}

\item{FUN2}{function to apply over each chunk of the abstract data frame}

\item{params}{a list of additional parameters}
}
\description{
Default method for applying a function to an abstract
data frame.
}
