% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.output.function.r
\name{default.output.function}
\alias{default.output.function}
\title{Default Method for Outputting ADF Calculations to Disk}
\usage{
default.output.function(x, FUN2, outDir, params)
}
\arguments{
\item{x}{an abstract data frame}

\item{FUN2}{function to apply over each chunk}

\item{outDir}{an empty directory for storing output}

\item{params}{a list of additional parameters}
}
\description{
Default method for constructing output file to file system.
}
